+++ app/lib/validations.js
@@ -9,8 +9,6 @@ export const isValidRepeatPassword = (walletPassword, repeatPassword) => walletP
 
 export const isNotEmptyString = (value) => value !== '';
 
-export const isValidCurrency = (value) => value === 'ada';
-
 export const isValidAmountInLovelaces = (value: string) => (
   isInt(value, {
     allow_leading_zeroes: false,