+++ app/stores/WalletsStore.js
@@ -1,5 +1,5 @@
 // @flow
-import { observable, computed, action, runInAction } from 'mobx';
+import { observable, computed, action, runInAction, untracked } from 'mobx';
 import _ from 'lodash';
 import Store from './lib/Store';
 import Wallet from '../domain/Wallet';
@@ -36,9 +36,8 @@ export default class WalletsStore extends Store {
   @observable restoreRequest: Request<RestoreWalletResponse> = new Request(this.api.restoreWallet);
   /* eslint-enable max-len */
 
-  _newWalletDetails: { name: string, currency: string, mnemonic: string, password: ?string, } = {
+  _newWalletDetails: { name: string, mnemonic: string, password: ?string, } = {
     name: '',
-    currency: '',
     mnemonic: '',
     password: null,
   };
@@ -55,7 +54,7 @@ export default class WalletsStore extends Store {
     walletBackup.finishWalletBackup.listen(this._finishWalletCreation);
     this.registerReactions([
       this._updateActiveWalletOnRouteChanges,
-      this._openAddWalletDialogWhenThereAreNoWallets,
+      this._toggleAddWalletDialogOnWalletsLoaded,
     ]);
     if (environment.CARDANO_API) {
       setInterval(this.refreshWalletsData, this.WALLET_REFRESH_INTERVAL);
@@ -64,7 +63,6 @@ export default class WalletsStore extends Store {
 
   _create = async (params: {
     name: string,
-    currency: string,
     password: ?string,
   }) => {
     Object.assign(this._newWalletDetails, params);
@@ -235,11 +233,11 @@ export default class WalletsStore extends Store {
     this.actions.router.goToRoute.trigger({ route });
   }
 
-  _openAddWalletDialogWhenThereAreNoWallets = () => {
+  _toggleAddWalletDialogOnWalletsLoaded = () => {
     if (this.hasLoadedWallets && !this.hasAnyWallets) {
-      this.actions.dialogs.open.trigger({
-        dialog: WalletAddDialog,
-      });
+      this.actions.dialogs.open.trigger({ dialog: WalletAddDialog });
+    } else if (untracked(() => this.stores.uiDialogs.isOpen(WalletAddDialog))) {
+      this.actions.dialogs.closeActiveDialog.trigger();
     }
   };